+++ packages/server/graphql/types/TeamMember.ts
@@ -11,7 +11,6 @@ import toTeamMemberId from 'parabol-client/utils/relay/toTeamMemberId'
 import {getUserId} from '../../utils/authorization'
 import {GQLContext} from '../graphql'
 import connectionFromTasks from '../queries/helpers/connectionFromTasks'
-import repoIntegrations from '../queries/repoIntegrations'
 import {resolveTeam} from '../resolvers'
 import GraphQLEmailType from './GraphQLEmailType'
 import GraphQLISO8601Type from './GraphQLISO8601Type'
@@ -94,7 +93,7 @@ const TeamMember = new GraphQLObjectType<any, GQLContext>({
       type: new GraphQLNonNull(GraphQLString),
       description: 'The name of the assignee'
     },
-    repoIntegrations,
+    repoIntegrations: require('../queries/repoIntegrations').default,
     tasks: {
       type: TaskConnection,
       description: 'Tasks owned by the team member',