+++ packages/server/postgres/queries/getAzureDevOpsDimensionFieldMaps.ts
@@ -11,15 +11,15 @@ const getAzureDevOpsDimensionFieldMaps = async (
   teamId: string,
   dimensionName: string,
   instanceId: string,
-  projectKey: string
+  projectKey: string,
+  workItemType: string
 ) => {
   // pg-typed doesnt' support records, so we can't use multiple composite keys
   // https://github.com/adelsz/pgtyped/issues/317
   const res = await getAzureDevOpsDimensionFieldMapsQuery.run(
-    {teamId, dimensionName, instanceId, projectKey},
+    {teamId, dimensionName, instanceId, projectKey, workItemType},
     getPg()
   )
-  const fieldMapEntry = res[0] as AzureDevOpsDimensionFieldMap
-  return fieldMapEntry
+  return res[0] as AzureDevOpsDimensionFieldMap
 }
 export default getAzureDevOpsDimensionFieldMaps