+++ packages/server/postgres/queries/src/getAzureDevOpsDimensionFieldMapsQuery.sql
@@ -2,4 +2,4 @@
   @name getAzureDevOpsDimensionFieldMapsQuery
 */
 SELECT * from "AzureDevOpsDimensionFieldMap"
-WHERE "teamId" = :teamId AND "dimensionName" = :dimensionName AND "instanceId" = :instanceId AND "projectKey" = :projectKey;
+WHERE "teamId" = :teamId AND "dimensionName" = :dimensionName AND "instanceId" = :instanceId AND "projectKey" = :projectKey AND "workItemType" = :workItemType;